funcionalidades de asientos agregadas

+++ com/asiento.go
@@ -5,5 +5,5 @@ type Asiento struct {
 	Zona      string
 	Fila      int16
 	Columna   int16
-	Estado    int16
+	Estado    int8
 }
